REFACTOR tabs & toolbar to optimize reuse
