+++ lib/components/src/bar/bar.js
@@ -1,4 +1,34 @@
+import React, { Children } from 'react';
+import PropTypes from 'prop-types';
 import { styled } from '@storybook/theming';
+
+const Side = styled.div(
+  {
+    display: 'flex',
+  },
+  ({ right }) =>
+    right
+      ? {
+          '& > *': {
+            marginRight: 15,
+          },
+        }
+      : {},
+  ({ left }) =>
+    left
+      ? {
+          '& > *': {
+            marginLeft: 15,
+          },
+        }
+      : {},
+  ({ flex }) =>
+    flex
+      ? {
+          flex: 1,
+        }
+      : {}
+);
 
 export const Bar = styled.div(
   ({ theme }) => ({
@@ -6,9 +36,13 @@ export const Bar = styled.div(
     justifyContent: 'space-between',
     background: theme.barFill,
     color: theme.barTextColor,
-    overflow: 'auto',
     height: 40,
     position: 'relative',
+  }),
+  ({ scroll = true }) => ({
+    overflow: scroll ? 'auto' : 'visible',
+    overflowX: scroll ? 'auto' : 'visible',
+    overflowY: scroll ? 'hidden' : 'visible',
   }),
   ({ theme, border }) =>
     border
@@ -19,3 +53,19 @@ export const Bar = styled.div(
         }
       : {}
 );
+
+export const FlexBar = ({ children, ...rest }) => {
+  const [left, right] = Children.toArray(children);
+  return (
+    <Bar {...rest}>
+      <Side flex left>
+        {left}
+      </Side>
+      {right ? <Side right>{right}</Side> : null}
+    </Bar>
+  );
+};
+
+FlexBar.propTypes = {
+  children: PropTypes.node.isRequired,
+};
