+++ lib/components/src/bar/separator.js
@@ -7,6 +7,7 @@ export const interleaveSeparators = list =>
       item ? (
         <Fragment key={item.id || item.key || `f-${index}`}>
           {acc}
+          {/* eslint-disable-next-line react/no-array-index-key */}
           {index > 0 ? <Separator key={`s-${index}`} /> : null}
           {item.render() || item}
         </Fragment>
@@ -23,9 +24,18 @@ export const Separator = styled.span(
     background: theme.mainBorderColor,
     marginTop: 8,
   }),
-  {
-    '& + &': {
-      display: 'none',
-    },
-  }
+  ({ force }) =>
+    force
+      ? {}
+      : {
+          '& + &': {
+            display: 'none',
+          },
+          '&:first-child': {
+            display: 'none',
+          },
+          '&:last-child': {
+            display: 'none',
+          },
+        }
 );
