+++ lib/components/src/index.js
@@ -12,13 +12,16 @@ export { default as Form } from './form/index';
 
 export SyntaxHighlighter from './syntaxhighlighter/syntaxhighlighter';
 
-export { Tabs, TabsState, TabWrapper, TabButton } from './tabs/tabs';
+export { Tabs, TabsState, TabWrapper } from './tabs/tabs';
 export { Popout, Item, Icon, Title, Detail, List, Badge } from './menu/menu';
 export { ActionBar, ActionButton } from './panel_actionbar/panel_actionbar';
 export { default as Placeholder } from './placeholder/placeholder';
 
 export { Explorer } from './explorer/explorer';
 export { Preview } from './preview/preview';
-export { IconButton, Separator, Toolbar } from './preview/toolbar';
+export { Toolbar } from './preview/toolbar';
+export { IconButton, TabButton } from './bar/button';
+export { Separator } from './bar/separator';
+export { Bar, FlexBar } from './bar/bar';
 
 export { default as Icons } from './icon/icon';
