+++ lib/components/src/preview/preview.js
@@ -2,25 +2,23 @@ import { window } from 'global';
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import memoize from 'memoizerific';
-import { styled } from '@storybook/theming';
 
 import { SET_CURRENT_STORY } from '@storybook/core-events';
 import { types } from '@storybook/addons';
 
 import Helmet from 'react-helmet-async';
 
-import { IconButton, Toolbar } from './toolbar';
+import { Toolbar } from './toolbar';
 import { interleaveSeparators } from '../bar/separator';
+import { IconButton, TabButton } from '../bar/button';
 
 import Icons from '../icon/icon';
-import { TabButton, TabBar } from '../tabs/tabs';
+import { TabBar } from '../tabs/tabs';
 
-import Zoom from './tools/zoom';
-import { Grid, Background } from './tools/background';
 import * as S from './components';
 
-import { ZoomProvider, ZoomConsumer } from './zoom';
-import { BackgroundProvider, BackgroundConsumer } from './background';
+import { ZoomProvider, ZoomConsumer, Zoom } from './zoom';
+import { Grid, Background, BackgroundProvider, BackgroundConsumer } from './background';
 
 import { IFrame } from './iframe';
 
@@ -75,17 +73,13 @@ const defaultWrappers = [
   },
 ];
 
-const Tabs = styled(TabBar)({
-  '&:first-child': { marginLeft: 0 },
-});
-
 const getTools = memoize(10)(
   (getElements, panels, actions, options, storyId, viewMode, location, path, baseUrl) => {
     const tools = getElementList(getElements, types.TOOL, [
       panels.filter(p => p.id !== 'canvas').length
         ? {
             render: () => (
-              <Tabs key="tabs" scroll={false}>
+              <TabBar key="tabs" scroll={false}>
                 {panels.map((t, index) => {
                   const to = t.route({ storyId, viewMode, path, location });
                   const isActive = t.match({ storyId, viewMode, path, location });
@@ -95,7 +89,7 @@ const getTools = memoize(10)(
                     </S.UnstyledLink>
                   );
                 })}
-              </Tabs>
+              </TabBar>
             ),
           }
         : null,
@@ -227,7 +221,10 @@ class Preview extends Component {
                 <title>{description ? `${description} ⋅ ` : ''}Storybook</title>
               </Helmet>
             )}
-            <Toolbar key="toolbar" shown={options.isToolshown} left={left} right={right} />
+            <Toolbar key="toolbar" shown={options.isToolshown} border>
+              <Fragment key="left">{left}</Fragment>
+              <Fragment key="right">{right}</Fragment>
+            </Toolbar>
             <S.FrameWrap key="frame" offset={toolbarHeight}>
               {panels.map(p => (
                 <Fragment key={p.id || p.key}>
