+++ lib/components/src/tabs/tabs.js
@@ -1,9 +1,10 @@
-import React, { Component } from 'react';
+import React, { Children, Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { styled } from '@storybook/theming';
 
 import Placeholder from '../placeholder/placeholder';
-import { Bar } from '../bar/bar';
+import { FlexBar } from '../bar/bar';
+import { TabButton } from '../bar/button';
 
 const Wrapper = styled.div(
   ({ theme, bordered }) =>
@@ -30,79 +31,26 @@ const Wrapper = styled.div(
 );
 
 export const TabBar = styled.div(
-  ({ scroll }) => ({
-    overflow: scroll ? 'auto' : 'visisble',
+  {
     whiteSpace: 'nowrap',
     height: '100%',
+
+    '&:first-child': {
+      marginLeft: 0,
+    },
+  },
+  ({ scroll }) => ({
+    // super lame, hiding of scrollbar
     paddingBottom: 50,
+    overflowX: scroll ? 'auto' : 'visisble',
+    overflowY: scroll ? 'hidden' : 'visisble',
   }),
   ({ flex }) =>
     flex
       ? {
-          flex: 1,
+          flex: typeof flex === 'number' ? flex : 1,
         }
       : {}
-);
-
-const TabBarTools = styled.div({
-  height: '100%',
-  paddingRight: 15,
-  paddingLeft: 15,
-  display: 'flex',
-  alignItems: 'center',
-
-  '& > *': {
-    marginRight: 15,
-  },
-  '& > *:last-child': {
-    marginRight: 0,
-  },
-});
-
-export const TabButton = styled.button(
-  {
-    whiteSpace: 'normal',
-    display: 'inline-flex',
-    overflow: 'hidden',
-    verticalAlign: 'top',
-    justifyContent: 'center',
-    alignItems: 'center',
-    textAlign: 'center',
-    textDecoration: 'none',
-
-    '&:empty': {
-      display: 'none',
-    },
-  },
-  ({ theme }) => ({
-    padding: '0 15px',
-    textTransform: 'capitalize',
-    transition: 'color 0.2s linear, border-bottom-color 0.2s linear',
-    height: 40,
-    lineHeight: '12px',
-    boxSizing: 'border-box',
-    cursor: 'pointer',
-    background: 'transparent',
-    border: '0 solid transparent',
-    borderTop: '3px solid transparent',
-    borderBottom: '3px solid transparent',
-    fontWeight: 'bold',
-
-    '&:focus': {
-      outline: '0 none',
-      borderBottomColor: theme.colors.highlight,
-    },
-  }),
-  ({ active, theme }) =>
-    active
-      ? {
-          color: theme.barSelectedColor,
-          borderBottomColor: theme.barSelectedColor,
-        }
-      : {
-          color: theme.dimmedTextColor,
-          borderBottomColor: 'transparent',
-        }
 );
 
 const Content = styled.div(
@@ -120,26 +68,6 @@ const Content = styled.div(
         }
       : {}
 );
-
-export const Tab = ({ active, name, title, onSelect }) => {
-  const onClick = e => {
-    e.preventDefault();
-    onSelect(name);
-  };
-
-  return (
-    <TabButton type="button" key={name} active={active} onClick={onClick} role="tab">
-      {typeof title === 'function' ? title() : title}
-    </TabButton>
-  );
-};
-
-Tab.propTypes = {
-  active: PropTypes.bool.isRequired,
-  name: PropTypes.string.isRequired,
-  title: PropTypes.oneOfType([PropTypes.func, PropTypes.node]).isRequired,
-  onSelect: PropTypes.func.isRequired,
-};
 
 const VisuallyHidden = styled.div(({ active }) =>
   active ? { display: 'block' } : { display: 'none' }
@@ -161,46 +89,47 @@ export const panelProps = {
   active: PropTypes.bool,
 };
 
+const childrenToList = (children, selected) =>
+  Children.toArray(children).map(({ props: { title, id, children: childrenOfChild } }, index) => {
+    const content = Array.isArray(childrenOfChild) ? childrenOfChild[0] : childrenOfChild;
+    return {
+      active: selected ? id === selected : index === 0,
+      title,
+      id,
+      render:
+        typeof content === 'function'
+          ? content
+          : // eslint-disable-next-line react/prop-types
+            ({ active, key }) => (
+              <VisuallyHidden key={key} active={active} role="tabpanel">
+                {content}
+              </VisuallyHidden>
+            ),
+    };
+  });
+
 export const Tabs = React.memo(
-  ({ children, selected, actions, absolute, bordered, scroll, tools, flex, id: htmlId }) => {
-    const list = React.Children.toArray(children).map(
-      ({ props: { title, id, children: childrenOfChild } }, index) => {
-        const content = Array.isArray(childrenOfChild) ? childrenOfChild[0] : childrenOfChild;
-        return {
-          active: selected ? id === selected : index === 0,
-          title,
-          id,
-          render:
-            typeof content === 'function'
-              ? content
-              : // eslint-disable-next-line react/prop-types
-                ({ active, key }) => (
-                  <VisuallyHidden key={key} active={active} role="tabpanel">
-                    {content}
-                  </VisuallyHidden>
-                ),
-        };
-      }
-    );
+  ({ children, selected, actions, absolute, bordered, scroll, tools, id: htmlId }) => {
+    const list = childrenToList(children, selected);
 
     return list.length ? (
       <Wrapper absolute={absolute} bordered={bordered} id={htmlId}>
-        <Bar border>
-          <TabBar role="tablist" scroll={scroll} flex={flex}>
+        <FlexBar border scroll={!scroll}>
+          <TabBar role="tablist">
             {list.map(({ title, id, active }) => (
-              <Tab
-                key={`${id}-bar`}
+              <TabButton
+                type="button"
+                key={id}
                 active={active}
-                name={id}
-                title={title}
-                onSelect={() => {
-                  actions.onSelect(id);
-                }}
-              />
+                onClick={e => e.preventDefault() || actions.onSelect(id)}
+                role="tab"
+              >
+                {typeof title === 'function' ? title() : title}
+              </TabButton>
             ))}
           </TabBar>
-          {tools ? <TabBarTools>{tools}</TabBarTools> : null}
-        </Bar>
+          {tools ? <Fragment>{tools}</Fragment> : null}
+        </FlexBar>
         <Content absolute={absolute}>
           {list.map(({ id, active, render }) => render({ key: id, active }))}
         </Content>
@@ -219,7 +148,6 @@ Tabs.propTypes = {
   actions: PropTypes.shape({
     onSelect: PropTypes.func.isRequired,
   }).isRequired,
-  flex: PropTypes.bool,
   absolute: PropTypes.bool,
   bordered: PropTypes.bool,
   scroll: PropTypes.bool,
@@ -231,7 +159,6 @@ Tabs.defaultProps = {
   selected: null,
   absolute: false,
   bordered: false,
-  flex: false,
   scroll: true,
 };
