+++ lib/ui/src/components/panel/panel.js
@@ -1,7 +1,6 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
-import { styled } from '@storybook/theming';
-import { Tabs, Icons } from '@storybook/components';
+import { Tabs, Icons, IconButton, Separator } from '@storybook/components';
 
 class SafeTab extends Component {
   constructor(props) {
@@ -37,25 +36,6 @@ SafeTab.defaultProps = {
   children: null,
 };
 
-const ToolButton = styled.button({
-  background: 'none',
-  border: '0 none',
-  padding: 0,
-  color: 'inherit',
-
-  '&:hover, &:focus': {
-    color: '#300aaff',
-    outline: '0 none',
-    cursor: 'pointer',
-  },
-});
-
-export const Separator = styled.span({
-  width: 1,
-  height: 24,
-  background: '#eee',
-});
-
 // eslint-disable-next-line react/no-multi-comp
 const AddonPanel = React.memo(({ panels, actions, selectedPanel, panelPosition }) => (
   <Tabs
@@ -65,13 +45,13 @@ const AddonPanel = React.memo(({ panels, actions, selectedPanel, panelPosition }
     flex
     tools={
       <Fragment>
-        <Separator key="1" />
-        <ToolButton key="position" onClick={actions.togglePosition}>
+        <Separator force key="1" />
+        <IconButton key="position" onClick={actions.togglePosition}>
           <Icons icon={panelPosition === 'bottom' ? 'bottombar' : 'sidebaralt'} />
-        </ToolButton>
-        <ToolButton key="visibility" onClick={actions.toggleVisibility}>
+        </IconButton>
+        <IconButton key="visibility" onClick={actions.toggleVisibility}>
           <Icons icon="close" />
-        </ToolButton>
+        </IconButton>
       </Fragment>
     }
     id="storybook-panel-root"
